+++ src/parser.rs
@@ -22,17 +22,24 @@ use super::tokenizer::*;
 use std::error::Error;
 use std::fmt;
 
-#[derive(Debug, Clone, PartialEq)]
-pub enum ParserError {
-    TokenizerError(String),
-    ParserError(String),
-}
+use crate::ast::{ParsedDate, ParsedTimestamp};
 
 // Use `Parser::expected` instead, if possible
 macro_rules! parser_err {
     ($MSG:expr) => {
         Err(ParserError::ParserError($MSG.to_string()))
     };
+    ($($arg:tt)*) => {
+        Err(ParserError::ParserError(format!($($arg)*)))
+    };
+}
+
+mod datetime;
+
+#[derive(Debug, Clone, PartialEq)]
+pub enum ParserError {
+    TokenizerError(String),
+    ParserError(String),
 }
 
 #[derive(PartialEq)]
@@ -196,7 +203,7 @@ impl Parser {
                 }
                 "CASE" => self.parse_case_expr(),
                 "CAST" => self.parse_cast_expr(),
-                "DATE" => Ok(Expr::Value(Value::Date(self.parse_literal_string()?))),
+                "DATE" => Ok(Expr::Value(self.parse_date()?)),
                 "EXISTS" => self.parse_exists_expr(),
                 "EXTRACT" => self.parse_extract_expr(),
                 "INTERVAL" => self.parse_literal_interval(),
@@ -205,7 +212,7 @@ impl Parser {
                     expr: Box::new(self.parse_subexpr(Self::UNARY_NOT_PREC)?),
                 }),
                 "TIME" => Ok(Expr::Value(Value::Time(self.parse_literal_string()?))),
-                "TIMESTAMP" => Ok(Expr::Value(Value::Timestamp(self.parse_literal_string()?))),
+                "TIMESTAMP" => Ok(Expr::Value(self.parse_timestamp()?)),
                 // Here `w` is a word, check if it's a part of a multi-part
                 // identifier, a function call, or a simple identifier:
                 _ => match self.peek_token() {
@@ -453,6 +460,117 @@ impl Parser {
         }
     }
 
+    fn parse_date(&mut self) -> Result<Value, ParserError> {
+        use std::convert::TryInto;
+
+        let value = self.parse_literal_string()?;
+        let pdt = Self::parse_interval_string(&value, &DateTimeField::Year)?;
+
+        match (pdt.year, pdt.month, pdt.day, pdt.hour) {
+            (Some(year), Some(month), Some(day), None) => {
+                let p_err = |e: std::num::TryFromIntError, field: &str| {
+                    ParserError::ParserError(format!(
+                        "{} in date '{}' is invalid: {}",
+                        field, value, e
+                    ))
+                };
+
+                // type inference with try_into() fails so we need to mutate it negative
+                let mut year: i64 = year.try_into().map_err(|e| p_err(e, "Year"))?;
+                year *= pdt.positivity();
+                if month > 12 || month == 0 {
+                    return parser_err!(
+                        "Month in date '{}' must be a number between 1 and 12, got: {}",
+                        value,
+                        month
+                    );
+                }
+                let month: u8 = month.try_into().expect("invalid month");
+                let day: u8 = day.try_into().map_err(|e| p_err(e, "Day"))?;
+                if day == 0 {
+                    parser_err!("Day in date '{}' cannot be zero: {}", value, day)?;
+                }
+                Ok(Value::Date(value, ParsedDate { year, month, day }))
+            }
+            (Some(_), Some(_), Some(_), Some(hours)) => parser_err!(
+                "Hours cannot be supplied for DATE, got {} in '{}'",
+                hours,
+                value
+            ),
+            (_, _, _, _) => Err(ParserError::ParserError(format!(
+                "year, day and month are all required, got: '{}'",
+                value
+            ))),
+        }
+    }
+
+    fn parse_timestamp(&mut self) -> Result<Value, ParserError> {
+        use std::convert::TryInto;
+
+        let value = self.parse_literal_string()?;
+        let pdt = Self::parse_interval_string(&value, &DateTimeField::Year)?;
+
+        match (
+            pdt.year, pdt.month, pdt.day, pdt.hour, pdt.minute, pdt.second, pdt.nano,
+        ) {
+            (Some(year), Some(month), Some(day), Some(hour), Some(minute), Some(second), nano) => {
+                let p_err = |e: std::num::TryFromIntError, field: &str| {
+                    ParserError::ParserError(format!(
+                        "{} in date '{}' is invalid: {}",
+                        field, value, e
+                    ))
+                };
+
+                // type inference with try_into() fails so we need to mutate it negative
+                let mut year: i64 = year.try_into().map_err(|e| p_err(e, "Year"))?;
+                year *= pdt.positivity();
+                if month > 12 || month == 0 {
+                    return parser_err!(
+                        "Month in date '{}' must be a number between 1 and 12, got: {}",
+                        value,
+                        month
+                    );
+                }
+                let month: u8 = month.try_into().expect("invalid month");
+                if month == 0 {
+                    parser_err!("Month in timestamp '{}' cannot be zero: {}", value, day)?;
+                }
+                let day: u8 = day.try_into().map_err(|e| p_err(e, "Day"))?;
+                if day == 0 {
+                    parser_err!("Day in timestamp '{}' cannot be zero: {}", value, day)?;
+                }
+                let hour: u8 = hour.try_into().map_err(|e| p_err(e, "Hour"))?;
+                if hour > 23 {
+                    parser_err!("Hour in timestamp '{}' cannot be > 23: {}", value, hour)?;
+                }
+                let minute: u8 = minute.try_into().map_err(|e| p_err(e, "Minute"))?;
+                if minute > 59 {
+                    parser_err!("Minute in timestamp '{}' cannot be > 59: {}", value, minute)?;
+                }
+                let second: u8 = second.try_into().map_err(|e| p_err(e, "Minute"))?;
+                if second > 60 {
+                    parser_err!("Second in timestamp '{}' cannot be > 60: {}", value, second)?;
+                }
+                Ok(Value::Timestamp(
+                    value,
+                    ParsedTimestamp {
+                        year,
+                        month,
+                        day,
+                        hour,
+                        minute,
+                        second,
+                        nano: nano.unwrap_or(0),
+                    },
+                ))
+            }
+            _ => Err(ParserError::ParserError(format!(
+                "timestamp is missing fields, year through second are all required, got: '{}'",
+                value
+            ))),
+        }
+    }
+
     /// Parse an INTERVAL literal.
     ///
     /// Some syntactically valid intervals:
@@ -466,14 +584,14 @@ impl Parser {
     ///
     /// Note that we do not currently attempt to parse the quoted value.
     pub fn parse_literal_interval(&mut self) -> Result<Expr, ParserError> {
-        // The SQL standard allows an optional sign before the value string, but
+        // The SQL standard allows an optional sign before the raw_value string, but
         // it is not clear if any implementations support that syntax, so we
         // don't currently try to parse it. (The sign can instead be included
-        // inside the value string.)
+        // inside the raw_value string.)
 
         // The first token in an interval is a string literal which specifies
         // the duration of the interval.
-        let value = self.parse_literal_string()?;
+        let raw_value = self.parse_literal_string()?;
 
         // Following the string literal is a qualifier which indicates the units
         // of the duration specified in the string literal.
@@ -507,13 +625,16 @@ impl Parser {
                 }
             };
 
-        Ok(Expr::Value(Value::Interval {
-            value,
+        let value = Self::parse_interval_string(&raw_value, &leading_field)?;
+
+        Ok(Expr::Value(Value::Interval(IntervalValue {
+            value: raw_value,
+            parsed: value,
             leading_field,
             leading_precision,
             last_field,
             fractional_seconds_precision: fsec_precision,
-        }))
+        })))
     }
 
     /// Parse an operator following an expression
@@ -586,6 +707,38 @@ impl Parser {
             // Can only happen if `get_next_precedence` got out of sync with this function
             panic!("No infix parser for token {:?}", tok)
         }
+    }
+
+    /// parse
+    ///
+    /// ```text
+    /// <unquoted interval string> ::=
+    ///   [ <sign> ] { <year-month literal> | <day-time literal> }
+    /// <year-month literal> ::=
+    ///     <years value> [ <minus sign> <months value> ]
+    ///   | <months value>
+    /// <day-time literal> ::=
+    ///     <day-time interval>
+    ///   | <time interval>
+    /// <day-time interval> ::=
+    ///   <days value> [ <space> <hours value> [ <colon> <minutes value>
+    ///       [ <colon> <seconds value> ] ] ]
+    /// <time interval> ::=
+    ///     <hours value> [ <colon> <minutes value> [ <colon> <seconds value> ] ]
+    ///   | <minutes value> [ <colon> <seconds value> ]
+    ///   | <seconds value>
+    /// ```
+    pub fn parse_interval_string(
+        value: &str,
+        leading_field: &DateTimeField,
+    ) -> Result<ParsedDateTime, ParserError> {
+        if value.is_empty() {
+            return Err(ParserError::ParserError(
+                "Interval date string is empty!".to_string(),
+            ));
+        }
+        let toks = datetime::tokenize_interval(value)?;
+        datetime::build_parsed_datetime(&toks, leading_field, value)
     }
 
     /// Parses the parens following the `[ NOT ] IN` operator